move set_cursor_visible
set_cursor_visible이 console 내용을 필요로 하기 때문에 console.c로 이동
