+++ console.c
@@ -21,3 +21,20 @@ void console_set_color(windows_console_t* console, uint8_t fore_color, uint8_t b
 void console_set_default_color(windows_console_t* console) {
     console_set_color(console, WHITE, BLACK);
 }
+
+
+//cursor
+void set_cursor_visible(windows_console_t* console, bool hide) {
+    console->cursor_info.is_hide = hide;
+
+    HANDLE hConsole = GetStdHandle(STD_OUTPUT_HANDLE); // 콘솔 핸들 가져오기
+    CONSOLE_CURSOR_INFO cursorInfo;
+
+    if (GetConsoleCursorInfo(hConsole, &cursorInfo)) {
+        cursorInfo.bVisible = console->cursor_info.is_hide; // TRUE: 표시, FALSE: 숨김
+        SetConsoleCursorInfo(hConsole, &cursorInfo);
+    }
+    else {
+        printf("Error: Unable to set console cursor visibility.\n");
+    }
+}
