+++ cursor.c
@@ -1,0 +1,42 @@
+#include "cursor.h"
+
+//cursor
+// 커서의 숨김,표시 정보를 가져오는 함수
+bool get_cursor_visibility(void) {
+    bool result = true; // tre가 default
+    HANDLE hConsole = GetStdHandle(STD_OUTPUT_HANDLE); // 콘솔 핸들 가져오기
+    CONSOLE_CURSOR_INFO cursorInfo;
+
+    // 현재 커서 정보 가져오기
+    if (GetConsoleCursorInfo(hConsole, &cursorInfo)) {
+        //cursorInfo.bVisible = visible; // TRUE: 표시, FALSE: 숨김
+        //SetConsoleCursorInfo(hConsole, &cursorInfo);
+        result = cursorInfo.bVisible;
+    }
+    else {
+        printf("Error: Unable to set console cursor visibility.\n");
+    }
+    return result;
+}
+
+
+
+//cursor
+void SetCurrentCursorPos(int x, int y) {
+    COORD pos; // = (x, y);
+    pos.X = x;
+    pos.Y = y;
+    SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), pos);
+}
+
+//cursor
+point_t GetCurrentCursorPos(void) {
+    point_t curr_point; // curPoint 타입이 운영체제 종속적이라서  point_t 타입을 만들어 이식성을 좋게 하려고 했다네요.
+    CONSOLE_SCREEN_BUFFER_INFO curr_info;
+
+    GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), &curr_info);
+    curr_point.x = curr_info.dwCursorPosition.X;
+    curr_point.y = curr_info.dwCursorPosition.Y;
+
+    return curr_point;
+}
